% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clear_All.R
\name{clear_all}
\alias{clear_all}
\title{Wyczyść środowisko i konsolę}
\usage{
clear_all()
}
\value{
NULL. Funkcja działa poprzez efekt uboczny na środowisko.
}
\description{
Funkcja usuwa wszystkie obiekty z globalnego środowiska,
odładowuje wszystkie niebazowe pakiety, czyści konsolę
oraz uruchamia garbage collector, aby zwolnić pamięć.
}
\examples{
clear_all()

}
