% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clip_Range.R
\name{clip_range}
\alias{clip_range}
\title{Przycinanie wartości do zadanego zakresu (dplyr)}
\usage{
clip_range(df, cols, min_val, max_val, in_place = TRUE)
}
\arguments{
\item{df}{Ramka danych zawierająca kolumny do przetworzenia.}

\item{cols}{Wektor znaków określający nazwy kolumn do oczyszczenia.}

\item{min_val}{Wartość minimalna – wszystkie wartości mniejsze zostaną zastąpione `NA`.}

\item{max_val}{Wartość maksymalna – wszystkie wartości większe zostaną zastąpione `NA`.}

\item{in_place}{Logiczne. Jeśli `TRUE`, nadpisuje oryginalne kolumny. 
Jeśli `FALSE`, tworzy nowe kolumny z przyrostkiem "_clipped".}
}
\value{
Ramka danych z oczyszczonymi kolumnami.
}
\description{
Funkcja zastępuje wszystkie wartości w wybranych kolumnach ramki danych, 
które są mniejsze od podanej wartości minimalnej lub większe od wartości maksymalnej, na `NA`. 
Można nadpisać oryginalne kolumny lub utworzyć nowe kolumny z oczyszczonymi wartościami.
}
\examples{
library(dplyr)
df <- data.frame(a = c(1, 5, 10, 15, 20),
                 b = c(-5, 0, 5, 10, 15))
clip_range(df, c("a", "b"), min_val = 0, max_val = 15, in_place = TRUE)
clip_range(df, c("a", "b"), min_val = 0, max_val = 15, in_place = FALSE)
}
