% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Raport.R
\name{data_report}
\alias{data_report}
\title{Generuje raport HTML z eksploracji danych}
\usage{
data_report(
  df,
  file = "data_report.html",
  title = "Raport danych",
  open = FALSE
)
}
\arguments{
\item{df}{Ramka danych do analizy.}

\item{file}{Nazwa pliku wyjściowego HTML (domyślnie "data_report.html").}

\item{title}{Tytuł raportu (domyślnie "Raport danych").}

\item{open}{Czy otworzyć raport w przeglądarce po wygenerowaniu (domyślnie FALSE).}
}
\value{
NULL. Funkcja zapisuje raport do pliku i opcjonalnie otwiera go w przeglądarce.
}
\description{
Funkcja tworzy automatyczny raport HTML z podstawową analizą ramki danych.
Raport zawiera m.in. wymiary danych, strukturę, brakujące wartości, liczbę zer,
typy zmiennych, liczbę unikalnych wartości, najczęstsze i najrzadsze wartości,
podstawowe statystyki kolumn numerycznych, korelacje oraz liczbę kategorii
w kolumnach faktorowych i znakowych.
}
\examples{
data_report(mtcars, file = "mtcars_report.html", title = "Raport mtcars")

}
