% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Visualization.R
\name{data_visualization}
\alias{data_visualization}
\title{Tworzy raport HTML z wizualizacjami danych}
\usage{
data_visualization(
  data,
  qualitative_cols = NULL,
  file = "plots.html",
  open = FALSE,
  css_file = "data_visualization.css"
)
}
\arguments{
\item{data}{Ramka danych do wizualizacji.}

\item{qualitative_cols}{Wektor nazw kolumn jakościowych (faktorowych lub character),
które mają być wyróżnione kolorami w wykresach (domyślnie NULL).}

\item{file}{Nazwa pliku wyjściowego HTML (domyślnie "plots.html").}

\item{open}{Czy otworzyć raport w przeglądarce po wygenerowaniu (domyślnie FALSE).}

\item{css_file}{Nazwa pliku CSS używanego w raporcie (domyślnie "data_visualization.css").}
}
\value{
NULL. Funkcja zapisuje raport do pliku HTML i opcjonalnie otwiera go w przeglądarce.
}
\description{
Funkcja generuje automatyczny raport HTML z wizualizacjami danych,
obejmujący m.in. wykresy typu GGally ggpairs, scatter, histogram,
density, violin, QQ oraz macierze korelacji. Dodatkowo wspiera
wyróżnianie kolumn jakościowych (faktorowych) poprzez kolorowanie.
}
\examples{
data_visualization(mtcars, qualitative_cols = c("cyl", "gear"), file = "mtcars_plots.html")

}
