% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Visualization_Simple.R
\name{data_visualization_simple}
\alias{data_visualization_simple}
\title{Tworzy raport HTML w formie flexdashboard z wizualizacjami danych}
\usage{
data_visualization_simple(
  data,
  qualitative_cols = NULL,
  index = NULL,
  file = "plots.html",
  author = NULL,
  location = NULL,
  title = "Data Visualization",
  open = FALSE
)
}
\arguments{
\item{data}{Ramka danych do wizualizacji.}

\item{qualitative_cols}{Wektor nazw kolumn jakościowych, które mają być wyróżnione kolorami (domyślnie NULL).}

\item{index}{Nazwa kolumny, która będzie używana jako oś X (np. czas). Jeśli NULL,
funkcja użyje kolejnych numerów wierszy.}

\item{file}{Nazwa pliku wyjściowego HTML (domyślnie "plots.html").}

\item{author}{Autor raportu (domyślnie NULL).}

\item{location}{Lokalizacja raportu (domyślnie NULL).}

\item{title}{Tytuł raportu (domyślnie "Data Visualization").}

\item{open}{Czy otworzyć raport w przeglądarce po wygenerowaniu (domyślnie FALSE).}
}
\value{
NULL. Funkcja zapisuje raport do pliku HTML i opcjonalnie otwiera go w przeglądarce.
}
\description{
Funkcja generuje automatyczny raport HTML w formie dashboardu, zawierający:
- wykresy typu histogram, density, violin, QQ,
- macierze korelacji (Pearson, Spearman, Kendall),
- informacje o sesji.
}
\details{
Wspiera wyróżnianie kolumn jakościowych (faktorowych lub character) poprzez kolorowanie.
}
\examples{
data_visualization_simple(iris, qualitative_cols = "Species", author = "John Doe", location = "Kraków", open = TRUE)

}
