% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date_Convert.R
\name{date_convert}
\alias{date_convert}
\title{Konwertuje kolumny z datami na obiekt klasy Date}
\usage{
date_convert(df, cols)
}
\arguments{
\item{df}{Ramka danych zawierająca kolumny do konwersji.}

\item{cols}{Wektor nazw kolumn, które mają zostać skonwertowane.}
}
\value{
Ramka danych z kolumnami zamienionymi na typ \code{Date}.
}
\description{
Funkcja automatycznie konwertuje wskazane kolumny w ramce danych na typ \code{Date}.
Obsługiwane są różne formaty dat, m.in. rok-miesiąc-dzień, dzień.miesiąc.rok itp.
Jeśli kolumna zawiera tylko rok i miesiąc, dodawany jest pierwszy dzień miesiąca.
}
\examples{
df <- data.frame(date1 = c("2023.06", "2023.07"), date2 = c("01-06-2023", "15-07-2023"))
date_convert(df, c("date1", "date2"))
}
