% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color.R
\name{get_color}
\alias{get_color}
\title{Zwraca kolor z podstawowej palety lub losowy kolor}
\usage{
get_color(n)
}
\arguments{
\item{n}{Liczba całkowita >= 1 określająca indeks koloru.}
}
\value{
String z nazwą koloru (np. "red") lub kodem HEX (np. "#A1B2C3").
}
\description{
Funkcja zwraca kolor w postaci nazwy lub kodu HEX. 
Jeśli podana liczba `n` mieści się w zakresie podstawowej palety,
zwracany jest kolor z tej palety. Jeśli `n` jest większe od długości
palety, generowany jest losowy kolor w formacie HEX.
}
\examples{
get_color(3)    # zwraca "green"
get_color(15)   # zwraca losowy kolor HEX

}
