% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/How_To_Use.R
\name{how_to_use}
\alias{how_to_use}
\title{Instrukcja obsługi pakietu}
\usage{
how_to_use()
}
\value{
Nic, wypisuje informacje na konsolę
}
\description{
Funkcja wypisuje na konsolę podstawowe informacje o pakiecie i kolejność użycia funkcji.
}
