% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Num_To_Qual.R
\name{num_to_qual}
\alias{num_to_qual}
\title{Zamienia kolumny numeryczne na jakościowe z automatycznymi etykietami lub własnymi labels}
\usage{
num_to_qual(
  df,
  cols = NULL,
  breaks = NULL,
  labels = NULL,
  include.lowest = TRUE,
  prefix = "cat_",
  replace = TRUE
)
}
\arguments{
\item{df}{Ramka danych.}

\item{cols}{Nazwy kolumn numerycznych do przekształcenia;
jeśli NULL — wszystkie kolumny numeryczne.}

\item{breaks}{Liczba przedziałów lub wektor granic (jak w `cut()`).}

\item{labels}{(opcjonalnie) Własne etykiety dla przedziałów.}

\item{include.lowest}{logiczne; czy pierwszy przedział zawiera wartości minimalne (domyślnie TRUE).}

\item{prefix}{Prefiks nazw nowych kolumn (jeśli `replace = FALSE`).}

\item{replace}{logiczne; czy zastąpić oryginalne kolumny (TRUE) czy dodać nowe (FALSE).}
}
\value{
Ramka danych z kolumnami jakościowymi.
}
\description{
Funkcja przekształca wskazane kolumny numeryczne (lub wszystkie numeryczne)
w kategorie jakościowe według określonych przedziałów.
Jeśli `labels` = NULL, etykiety są automatycznie generowane jako "A (low–high)".
}
\examples{
# automatyczne etykiety A (low–high)
num_to_qual(iris, cols = c("Sepal.Length", "Petal.Length"), breaks = 3)

# własne etykiety
num_to_qual(iris, cols = "Sepal.Length", breaks = 3, labels = c("niski", "średni", "wysoki"))

}
