% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Acf.R
\name{plot_acf}
\alias{plot_acf}
\title{Tworzy interaktywny wykres ACF dla pojedynczej kolumny}
\usage{
plot_acf(df, column, max_lag = 20)
}
\arguments{
\item{df}{Ramka danych zawierająca kolumnę do analizy.}

\item{column}{Nazwa kolumny numerycznej, dla której generujemy wykres ACF.}

\item{max_lag}{Maksymalna liczba lagów do policzenia ACF (domyślnie 20).}
}
\value{
Obiekt \pkg{plotly} z wykresem ACF dla podanej kolumny.
}
\description{
Funkcja generuje wykres ACF (autocorrelation function) dla jednej kolumny
numerycznej w ramce danych. Wykres jest interaktywny dzięki pakietowi \pkg{plotly}.
Punkty istotne statystycznie (poza przedziałem ufności ±1.96/sqrt(n)) są wyróżnione innym kolorem.
}
\examples{
plot_acf(mtcars, "mpg", max_lag = 15)
}
