% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Acf_Grid.R
\name{plot_acf_grid}
\alias{plot_acf_grid}
\title{Tworzy interaktywne wykresy ACF dla wszystkich kolumn numerycznych}
\usage{
plot_acf_grid(df, max_lag = 20)
}
\arguments{
\item{df}{Ramka danych zawierająca kolumny numeryczne do analizy.}

\item{max_lag}{Maksymalna liczba lagów do policzenia ACF (domyślnie 20).}
}
\value{
Lista obiektów \pkg{plotly} z wykresami ACF dla każdej kolumny numerycznej.
}
\description{
Funkcja generuje wykresy ACF (autocorrelation function) dla wszystkich
kolumn numerycznych w ramce danych. Wykresy są interaktywne dzięki pakietowi \pkg{plotly}.
Punkty istotne statystycznie (poza przedziałem ufności ±1.96/sqrt(n)) są wyróżnione innym kolorem.
}
\examples{
plot_acf_grid(mtcars, max_lag = 15)

}
