% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Correlation_Matrix.R
\name{plot_correlation_matrix}
\alias{plot_correlation_matrix}
\title{Rysuje macierz korelacji dla kolumn numerycznych}
\usage{
plot_correlation_matrix(df, method = "pearson", title = "Correlation Matrix")
}
\arguments{
\item{df}{Ramka danych zawierająca kolumny numeryczne do analizy korelacji.}

\item{method}{Metoda obliczania korelacji: "pearson" (domyślnie), "spearman" lub "kendall".}

\item{title}{Tytuł wykresu (domyślnie "Macierz korelacji").}
}
\value{
NULL. Funkcja generuje wykres jako efekt uboczny.
}
\description{
Funkcja generuje wizualizację macierzy korelacji dla wszystkich kolumn numerycznych
w ramce danych przy użyciu pakietu \pkg{corrplot}. Wyświetlane są wartości współczynników
oraz kolorowe pola odzwierciedlające siłę korelacji.
}
\examples{
plot_correlation_matrix(mtcars, method = "pearson", title = "Korelacje mtcars")

}
