% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Density.R
\name{plot_density}
\alias{plot_density}
\title{Tworzy interaktywny wykres gęstości (density) dla pojedynczej kolumny}
\usage{
plot_density(df, column, factor_cols = NULL)
}
\arguments{
\item{df}{Ramka danych zawierająca kolumnę numeryczną do wizualizacji.}

\item{column}{Nazwa kolumny numerycznej, dla której generowany jest wykres.}

\item{factor_cols}{Wektor nazw kolumn jakościowych (faktorowych) do kolorowania (domyślnie NULL — wszystkie dane traktowane jako jedna grupa).}
}
\value{
Obiekt \pkg{plotly} z wykresem gęstości dla podanej kolumny.
}
\description{
Funkcja generuje interaktywny wykres gęstości dla jednej kolumny
numerycznej w ramce danych. Możliwe jest rozdzielenie według kolumn
jakościowych (faktorowych), które są używane do wypełnienia kolorów.
}
\examples{
plot_density(mtcars, column = "mpg", factor_cols = c("cyl"))
}
