% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Density_Grid.R
\name{plot_density_grid}
\alias{plot_density_grid}
\title{Tworzy siatkę wykresów gęstości (density) dla kolumn numerycznych}
\usage{
plot_density_grid(df, factor_cols = NULL)
}
\arguments{
\item{df}{Ramka danych zawierająca kolumny numeryczne do wizualizacji.}

\item{factor_cols}{Wektor nazw kolumn jakościowych (faktorowych) do kolorowania
wykresów (domyślnie NULL — wszystkie dane traktowane jako jedna grupa).}
}
\value{
Obiekt klasy \code{list} z listą subplotów \pkg{plotly}.
}
\description{
Funkcja generuje interaktywne wykresy gęstości dla wszystkich kolumn numerycznych
w ramce danych. Możliwe jest rozdzielenie według kolumn jakościowych (faktorowych),
które są używane do wypełnienia kolorów. Wykresy są zwracane jako lista subplotów
dla łatwego umieszczenia w RMarkdown.
}
\examples{
plot_density_grid(mtcars, factor_cols = c("cyl"))

}
