% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Hist_Grid.R
\name{plot_hist_grid}
\alias{plot_hist_grid}
\title{Tworzy siatkę histogramów dla kolumn numerycznych}
\usage{
plot_hist_grid(df, factor_cols = NULL, log_y = FALSE)
}
\arguments{
\item{df}{Ramka danych zawierająca kolumny numeryczne do wizualizacji.}

\item{factor_cols}{Wektor nazw kolumn jakościowych (faktorowych) do kolorowania
i facetingu histogramów (domyślnie NULL — wszystkie dane traktowane jako jedna grupa).}
}
\value{
Obiekt klasy \code{list} z listą subplotów \pkg{plotly}.
}
\description{
Funkcja generuje interaktywne histogramy dla wszystkich kolumn numerycznych
w ramce danych. Możliwe jest rozdzielenie według kolumn jakościowych (faktorowych),
które są używane do wypełnienia kolorów i facetingu. Histogramy są zwracane
jako lista subplotów dla łatwego umieszczenia w RMarkdown.
}
\examples{
plot_hist_grid(mtcars, factor_cols = c("cyl"))

}
