% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Hist_LogY.R
\name{plot_hist_logY}
\alias{plot_hist_logY}
\title{Tworzy interaktywny histogram z logarytmiczną skalą osi Y dla pojedynczej kolumny}
\usage{
plot_hist_logY(df, column, factor_cols = NULL)
}
\arguments{
\item{df}{Ramka danych zawierająca kolumnę numeryczną do wizualizacji.}

\item{column}{Nazwa kolumny numerycznej, dla której generowany jest histogram.}

\item{factor_cols}{Wektor nazw kolumn jakościowych (faktorowych) do kolorowania i facetingu (domyślnie NULL — wszystkie dane traktowane jako jedna grupa).}
}
\value{
Obiekt \pkg{plotly} z histogramem dla podanej kolumny.
}
\description{
Funkcja generuje interaktywny histogram dla jednej kolumny numerycznej
w ramce danych z logarytmiczną skalą osi Y. Możliwe jest rozdzielenie
według kolumn jakościowych (faktorowych), które są używane do wypełnienia
kolorów i facetingu.
}
\examples{
plot_hist_logY(mtcars, column = "mpg", factor_cols = c("cyl"))
}
