% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Line.R
\name{plot_line}
\alias{plot_line}
\title{Tworzy interaktywny wykres liniowy dla jednej kolumny}
\usage{
plot_line(df, column, index = NULL, factor_cols = NULL, abline = TRUE)
}
\arguments{
\item{df}{Ramka danych zawierająca kolumnę numeryczną do wizualizacji.}

\item{column}{Nazwa kolumny numerycznej do wizualizacji.}

\item{index}{Nazwa kolumny, która będzie używana jako oś X (np. czas). Jeśli NULL, użyje kolejnych numerów wierszy.}

\item{factor_cols}{Wektor nazw kolumn jakościowych (faktorowych) do kolorowania linii (domyślnie NULL — wszystkie dane traktowane jako jedna grupa).}

\item{abline}{Logical. Czy dodać linię trendu regresji liniowej (domyślnie TRUE).}
}
\value{
Obiekt \pkg{plotly} z wykresem liniowym dla podanej kolumny.
}
\description{
Funkcja generuje wykres liniowy dla jednej kolumny numerycznej w ramce danych.
Możliwe jest rozdzielenie według kolumn jakościowych (faktorowych), które są używane
do kolorowania linii. Opcjonalnie można dodać linię trendu (regresji liniowej) jako przerywaną linię.
}
\examples{
plot_line(mtcars, column = "hp", index = "mpg", factor_cols = c("cyl"), abline = TRUE)
}
