% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Line_Grid.R
\name{plot_line_grid}
\alias{plot_line_grid}
\title{Tworzy interaktywne wykresy liniowe dla wszystkich kolumn numerycznych}
\usage{
plot_line_grid(df, index = NULL, factor_cols = NULL, abline = TRUE)
}
\arguments{
\item{df}{Ramka danych zawierająca kolumny numeryczne do wizualizacji.}

\item{index}{Nazwa kolumny, która będzie używana jako oś X (np. czas). Jeśli NULL,
funkcja użyje kolejnych numerów wierszy.}

\item{factor_cols}{Wektor nazw kolumn jakościowych (faktorowych) do kolorowania linii
(domyślnie NULL — wszystkie dane traktowane jako jedna grupa).}

\item{abline}{Logical. Czy dodać linię trendu regresji liniowej (domyślnie TRUE).}
}
\value{
Lista obiektów \pkg{plotly} z wykresami liniowymi dla każdej kolumny numerycznej.
}
\description{
Funkcja generuje wykresy liniowe dla wszystkich kolumn numerycznych w ramce danych.
Możliwe jest rozdzielenie według kolumn jakościowych (faktorowych), które są używane
do kolorowania linii. Opcjonalnie można dodać linię trendu (regresji liniowej) jako przerywaną linię.
}
\examples{
plot_line_grid(mtcars, index = "mpg", factor_cols = c("cyl"), abline = TRUE)
plot_line_grid(mtcars, abline = FALSE)
}
