% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Pacf.R
\name{plot_pacf}
\alias{plot_pacf}
\title{Tworzy interaktywny wykres PACF dla jednej kolumny}
\usage{
plot_pacf(df, column, max_lag = 20)
}
\arguments{
\item{df}{Ramka danych zawierająca kolumnę numeryczną do analizy.}

\item{column}{Nazwa kolumny numerycznej do analizy.}

\item{max_lag}{Maksymalna liczba lagów do policzenia PACF (domyślnie 20).}
}
\value{
Obiekt \pkg{plotly} z wykresem PACF dla podanej kolumny.
}
\description{
Funkcja generuje wykres PACF (partial autocorrelation function) dla jednej
kolumny numerycznej w ramce danych. Punkty istotne statystycznie
(poza przedziałem ufności ±1.96/sqrt(n)) są wyróżnione innym kolorem.
}
\examples{
plot_pacf(mtcars, column = "mpg", max_lag = 15)
}
