% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_QQ.R
\name{plot_qq}
\alias{plot_qq}
\title{Tworzy interaktywny wykres Q-Q dla jednej kolumny}
\usage{
plot_qq(df, column, factor_col = NULL)
}
\arguments{
\item{df}{Ramka danych zawierająca kolumnę numeryczną do wizualizacji.}

\item{column}{Nazwa kolumny numerycznej do analizy.}

\item{factor_col}{Nazwa kolumny jakościowej (faktorowej) do kolorowania (domyślnie NULL).}
}
\value{
Obiekt \pkg{plotly} z wykresem Q-Q dla podanej kolumny.
}
\description{
Funkcja generuje wykres Q-Q (quantile-quantile) dla jednej kolumny numerycznej.
Możliwe jest rozdzielenie według jednej kolumny jakościowej (faktorowej), 
która jest używana do kolorowania punktów.
}
\examples{
plot_qq(mtcars, column = "mpg", factor_col = "cyl")
}
