% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_QQ_Grid.R
\name{plot_qq_grid}
\alias{plot_qq_grid}
\title{Tworzy siatkę wykresów Q-Q dla kolumn numerycznych}
\usage{
plot_qq_grid(df, factor_cols = NULL, palette = "jco")
}
\arguments{
\item{df}{Ramka danych zawierająca kolumny numeryczne do wizualizacji.}

\item{factor_cols}{Wektor nazw kolumn jakościowych (faktorowych) do kolorowania
wykresów (domyślnie NULL — wszystkie dane traktowane jako jedna grupa).}

\item{palette}{Paleta kolorów dla grup faktorowych (domyślnie "jco").}
}
\value{
Obiekt klasy \code{tagList} z listą subplotów \pkg{plotly}.
}
\description{
Funkcja generuje interaktywne wykresy Q-Q dla wszystkich kolumn numerycznych
w ramce danych. Możliwe jest rozdzielenie według kolumn jakościowych (faktorowych),
które są używane do kolorowania. Wykresy są zwracane jako lista subplotów
dla łatwego umieszczenia w RMarkdown.
}
\examples{
plot_qq_grid(mtcars, factor_cols = c("cyl"), palette = "jco")

}
