% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Tworzy interaktywny wykres punktowy (scatter) dla jednej pary kolumn numerycznych}
\usage{
plot_scatter(df, x_col, y_col, factor_col = NULL)
}
\arguments{
\item{df}{Ramka danych zawierająca kolumny numeryczne do wizualizacji.}

\item{x_col}{Nazwa kolumny numerycznej dla osi X.}

\item{y_col}{Nazwa kolumny numerycznej dla osi Y.}

\item{factor_col}{Nazwa kolumny jakościowej (faktorowej) do kolorowania punktów (domyślnie NULL).}
}
\value{
Obiekt \pkg{plotly} z wykresem punktowym.
}
\description{
Funkcja generuje wykres punktowy (scatter) dla jednej pary kolumn numerycznych.
Możliwe jest rozdzielenie według jednej kolumny jakościowej (faktorowej), która
jest używana do kolorowania punktów.
}
\examples{
plot_scatter(mtcars, x_col = "mpg", y_col = "disp", factor_col = "cyl")
}
