% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Scatter_Ellipse.R
\name{plot_scatter_ellipse}
\alias{plot_scatter_ellipse}
\title{Tworzy interaktywny wykres punktowy z elipsą dla jednej pary kolumn numerycznych}
\usage{
plot_scatter_ellipse(df, x_col, y_col, factor_col = NULL)
}
\arguments{
\item{df}{Ramka danych zawierająca kolumny numeryczne do wizualizacji.}

\item{x_col}{Nazwa kolumny numerycznej dla osi X.}

\item{y_col}{Nazwa kolumny numerycznej dla osi Y.}

\item{factor_col}{Nazwa kolumny jakościowej (faktorowej) do kolorowania punktów i elips (domyślnie NULL).}
}
\value{
Obiekt \pkg{plotly} z wykresem punktowym i elipsą.
}
\description{
Funkcja generuje wykres punktowy (scatter) dla jednej pary kolumn numerycznych.
Dodatkowo rysowana jest elipsa obejmująca 95% punktów dla każdej grupy faktorowej.
}
\examples{
plot_scatter_ellipse(mtcars, x_col = "mpg", y_col = "disp", factor_col = "cyl")
}
