% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Scatter_Grid.R
\name{plot_scatter_grid}
\alias{plot_scatter_grid}
\title{Tworzy siatkę wykresów punktowych (scatter) dla par kolumn numerycznych}
\usage{
plot_scatter_grid(df, factor_cols = NULL)
}
\arguments{
\item{df}{Ramka danych zawierająca kolumny numeryczne do wizualizacji.}

\item{factor_cols}{Wektor nazw kolumn jakościowych (faktorowych) do kolorowania
punktów (domyślnie NULL — wszystkie dane traktowane jako jedna grupa).}
}
\value{
Obiekt klasy \code{tagList} z listą subplotów \pkg{plotly}.
}
\description{
Funkcja generuje interaktywne wykresy punktowe dla wszystkich par kolumn
numerycznych w ramce danych. Możliwe jest rozdzielenie według kolumn jakościowych
(faktorowych), które są używane do kolorowania punktów. Wykresy są zwracane
jako lista subplotów dla łatwego umieszczenia w RMarkdown.
}
\examples{
plot_scatter_grid(mtcars, factor_cols = c("cyl"))

}
