% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Violin.R
\name{plot_violin}
\alias{plot_violin}
\title{Tworzy interaktywny wykres violin dla jednej kolumny numerycznej}
\usage{
plot_violin(df, numeric_col, factor_col = NULL)
}
\arguments{
\item{df}{Ramka danych zawierająca kolumnę numeryczną do wizualizacji.}

\item{numeric_col}{Nazwa kolumny numerycznej.}

\item{factor_col}{Nazwa kolumny jakościowej (faktorowej) do kolorowania i grupowania (domyślnie NULL).}
}
\value{
Obiekt \pkg{plotly} z wykresem violin.
}
\description{
Funkcja generuje wykres violin dla jednej kolumny numerycznej.
Możliwe jest rozdzielenie według jednej kolumny jakościowej (faktorowej),
która jest używana do kolorowania i grupowania. Dodatkowo wyświetlany jest
boxplot dla wizualizacji mediany i kwartylów.
}
\examples{
plot_violin(mtcars, numeric_col = "mpg", factor_col = "cyl")
}
