% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Violin_Grid.R
\name{plot_violin_grid}
\alias{plot_violin_grid}
\title{Tworzy siatkę wykresów violin dla kolumn numerycznych}
\usage{
plot_violin_grid(df, factor_cols = NULL)
}
\arguments{
\item{df}{Ramka danych zawierająca kolumny numeryczne do wizualizacji.}

\item{factor_cols}{Wektor nazw kolumn jakościowych (faktorowych) do kolorowania
i grupowania wykresów (domyślnie NULL — wszystkie dane traktowane jako jedna grupa).}
}
\value{
Obiekt klasy \code{tagList} z listą subplotów \pkg{plotly}.
}
\description{
Funkcja generuje interaktywne wykresy violin dla wszystkich kolumn numerycznych
w ramce danych. Możliwe jest rozdzielenie według kolumn jakościowych (faktorowych),
które są używane do kolorowania i grupowania. Każdy violin zawiera dodatkowo
wykres boxplot dla wizualizacji mediany i kwartylów. Wykresy są zwracane
jako lista subplotów dla łatwego umieszczenia w RMarkdown.
}
\examples{
plot_violin_grid(mtcars, factor_cols = c("cyl"))

}
