% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_Qualitive.R
\name{prepare_qualitative}
\alias{prepare_qualitative}
\title{Przygotowuje ramkę danych do wizualizacji}
\usage{
prepare_qualitative(df, qualitative_cols = NULL)
}
\arguments{
\item{df}{Ramka danych}

\item{qualitative_cols}{Wektor nazw kolumn jakościowych (faktor/character) lub NULL}
}
\value{
Lista z elementami:
  \item{df}{zmodyfikowana lokalnie ramka danych}
  \item{qualitative_cols}{przetworzony wektor kolumn jakościowych}
}
\description{
Funkcja dodaje kolumnę "All" jeśli nie podano kolumn jakościowych,
lub konwertuje istniejące kolumny jakościowe na faktory lokalnie.
}
