% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero.handle.R
\name{zero.handle}
\alias{zero.handle}
\title{Obsługuje zera w wybranych kolumnach}
\usage{
zero.handle(df, cols, method = c("omit", "NA"))
}
\arguments{
\item{df}{Ramka danych do przefiltrowania.}

\item{cols}{Wektor nazw kolumn, w których sprawdzane są wartości zerowe.}

\item{method}{Metoda obsługi zer: "omit" – usuwa wiersze z zerami,
"NA" – zamienia zera na NA (domyślnie "omit").}
}
\value{
Ramka danych po przetworzeniu.
}
\description{
Funkcja usuwa lub zamienia wartości 0 w wybranych kolumnach w ramce danych.
}
\examples{
df <- data.frame(a = c(1,0,3), b = c(0,2,3))
zero.handle(df, cols = c("a", "b"), method = "omit")
zero.handle(df, cols = c("a", "b"), method = "NA")
}
