% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero.omit.R
\name{zero.omit}
\alias{zero.omit}
\title{Usuwa wiersze z zerami w wybranych kolumnach (dplyr)}
\usage{
zero.omit(df, cols)
}
\arguments{
\item{df}{Ramka danych do przefiltrowania.}

\item{cols}{Wektor nazw kolumn, w których sprawdzane są wartości zerowe.}
}
\value{
Ramka danych bez wierszy zawierających 0 w wybranych kolumnach.
}
\description{
Funkcja usuwa wszystkie wiersze z ramki danych, w których w kolumnach
podanych w argumencie \code{cols} występuje wartość 0.
}
\examples{
df <- data.frame(a = c(1,0,3), b = c(0,2,3))
zero.omit(df, cols = c("a", "b"))

}
