% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_GGpairs.R
\name{plot_ggpairs}
\alias{plot_ggpairs}
\title{Tworzy macierz par (pairplot) dla kolumn numerycznych z rozróżnieniem według zmiennych jakościowych}
\usage{
plot_ggpairs(df, qualitative_cols = NULL)
}
\arguments{
\item{df}{Ramka danych zawierająca kolumny numeryczne do wizualizacji.}

\item{qualitative_cols}{Wektor nazw kolumn jakościowych (faktorowych) użytych do kolorowania wykresów.
Domyślnie \code{NULL}, wtedy wszystkie dane traktowane są jako jedna grupa.}
}
\value{
Obiekt klasy \code{tagList} zawierający wykresy \code{ggpairs}, gotowy do renderowania w HTML/RMarkdown.
}
\description{
Funkcja generuje wykresy macierzy par (\code{GGally::ggpairs}) dla wszystkich kolumn numerycznych
w ramce danych. Dla każdej kolumny jakościowej (\code{qualitative_cols}) tworzony jest osobny wykres,
w którym kolor punktów odpowiada kategoriom danej zmiennej.
}
\examples{
plot_ggpairs(mtcars, qualitative_cols = c("cyl", "gear"))

}
